+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - Support for custom fields. These fields allow you to run custom filter
   functions for anything that cannot be expressed with Flop filters.
 - Added `fetch`, `get`, `get_all`, `delete`, `delete_first`, `drop`, `new`,
-  `take`, `pop` and `pop_first` functions to `Flop.Filter`.
+  `take`, `pop`, `pop_first`, `put` and `put_new` functions to `Flop.Filter`.
 
 ## [0.18.4] - 2022-11-17